pyproject.toml
[tool.poetry]
name = "flapkap-vending-api"
version = "0.1.0"
description = "an API for a vending machine, allowing users with a “seller” role to add, update or remove products, while users with a “buyer” role"
authors = ["Mohamed Tarek <mohamed96tarek@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"
django-stubs = "^4.2.7"
python-dotenv = "^1.0.1"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
